Working on crud unittesting removed uneeded class

+++ src/main/java/com/patrick/madskills/repository/CategoryRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.patrick.madskills.domain.Category;
 
 
-public interface CategoryRepository extends JpaRepository<Category, Long> { }
+public interface CategoryRepository extends JpaRepository<Category, Integer> { }
